requirements.txt
bertopic==0.16.0
gensim==4.3.2
keyphrase_vectorizers==0.0.11
llama_index==0.10.20
matplotlib==3.4.3
numpy==1.24.4
pandas==2.0.3
plotly==5.15.0
python-dotenv==1.0.1
scikit_learn==1.3.2
scipy==1.6.2
sentence_transformers==2.5.0
spacy==3.5.4
tqdm==4.66.2
umap_learn==0.5.5
llama-index==0.10.17
gensim==4.3.2
python-dotenv==1.0.1
langchain==0.1.12
llama-index-embeddings-langchain==0.1.2
openpyxl==3.1.3   # Not entirely sure about this